Draft: split Array from Draft.py
